Refactor table for fileresources
