+++ dc-cudami-admin/src/main/java/de/digitalcollections/cudami/admin/controller/identifiable/resource/FileResourcesMetadataController.java
@@ -8,28 +8,18 @@ import de.digitalcollections.cudami.client.identifiable.resource.CudamiFileResou
 import de.digitalcollections.model.exception.ResourceNotFoundException;
 import de.digitalcollections.model.exception.TechnicalException;
 import de.digitalcollections.model.identifiable.resource.FileResource;
-import de.digitalcollections.model.list.paging.PageRequest;
-import de.digitalcollections.model.list.paging.PageResponse;
-import de.digitalcollections.model.list.sorting.Order;
-import de.digitalcollections.model.list.sorting.Sorting;
 import java.util.List;
 import java.util.Locale;
 import java.util.UUID;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.i18n.LocaleContextHolder;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 /** Controller for resource management pages. */
 @Controller
@@ -55,12 +45,6 @@ public class FileResourcesMetadataController extends AbstractController {
     return "fileresources/create";
   }
 
-  @GetMapping("/api/fileresources/new")
-  @ResponseBody
-  public FileResource create() throws TechnicalException {
-    return service.create();
-  }
-
   @GetMapping("/fileresources/{uuid}/edit")
   public String edit(
       @PathVariable UUID uuid,
@@ -84,28 +68,6 @@ public class FileResourcesMetadataController extends AbstractController {
     return "fileresources/edit";
   }
 
-  @GetMapping("/api/fileresources")
-  @ResponseBody
-  public PageResponse<FileResource> find(
-      @RequestParam(name = "pageNumber", required = false, defaultValue = "0") int pageNumber,
-      @RequestParam(name = "pageSize", required = false, defaultValue = "25") int pageSize,
-      @RequestParam(name = "searchTerm", required = false) String searchTerm,
-      @RequestParam(name = "sortBy", required = false) List<Order> sortBy)
-      throws TechnicalException {
-    PageRequest pageRequest = new PageRequest(searchTerm, pageNumber, pageSize);
-    if (sortBy != null) {
-      Sorting sorting = new Sorting(sortBy);
-      pageRequest.setSorting(sorting);
-    }
-    return service.find(pageRequest);
-  }
-
-  @GetMapping("/api/fileresources/{uuid}")
-  @ResponseBody
-  public FileResource getByUuid(@PathVariable UUID uuid) throws TechnicalException {
-    return service.getByUuid(uuid);
-  }
-
   @GetMapping("/fileresources")
   public String list(Model model) throws TechnicalException {
     final Locale displayLocale = LocaleContextHolder.getLocale();
@@ -118,45 +80,6 @@ public class FileResourcesMetadataController extends AbstractController {
   @ModelAttribute("menu")
   protected String module() {
     return "fileresources";
-  }
-
-  @PostMapping("/api/fileresources")
-  public ResponseEntity save(@RequestBody FileResource fileResource) {
-    try {
-      FileResource fileResourceDb = service.save(fileResource);
-      return ResponseEntity.status(HttpStatus.CREATED).body(fileResourceDb);
-    } catch (TechnicalException e) {
-      LOGGER.error("Cannot save fileresource: ", e);
-      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
-    }
-  }
-
-  @GetMapping("/api/fileresources/type/{type}")
-  @ResponseBody
-  public PageResponse<FileResource> searchFileResourcesByType(
-      @PathVariable String type,
-      @RequestParam(name = "pageNumber", required = false, defaultValue = "0") int pageNumber,
-      @RequestParam(name = "pageSize", required = false, defaultValue = "5") int pageSize,
-      @RequestParam(name = "searchTerm", required = false) String searchTerm,
-      @RequestParam(name = "sortBy", required = false) List<Order> sortBy)
-      throws TechnicalException {
-    PageRequest pageRequest = new PageRequest(searchTerm, pageNumber, pageSize);
-    if (sortBy != null) {
-      Sorting sorting = new Sorting(sortBy);
-      pageRequest.setSorting(sorting);
-    }
-    return service.findByType(pageRequest, type);
-  }
-
-  @PutMapping("/api/fileresources/{uuid}")
-  public ResponseEntity update(@PathVariable UUID uuid, @RequestBody FileResource fileResource) {
-    try {
-      FileResource fileResourceDb = service.update(uuid, fileResource);
-      return ResponseEntity.ok(fileResourceDb);
-    } catch (TechnicalException e) {
-      LOGGER.error("Cannot save fileresource with uuid={}", uuid, e);
-      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
-    }
   }
 
   @GetMapping(value = "/fileresources/{uuid}")
