+++ dc-cudami-admin/src/main/resources/messages.properties
@@ -117,12 +117,15 @@ lbl.collection.inline=collection
 lbl.created=created
 lbl.description=Abstract
 lbl.digital_object.inline=the digital object
+lbl.filename=Filename
+lbl.filesize=Filesize
 lbl.firstname=First name
 lbl.identifier_type=Identifier type
 lbl.itemlocale=Show language dependent values for
 lbl.label=Label
 lbl.last_modified=last modified
 lbl.lastname=Last name
+lbl.mimetype=Mimetype
 lbl.name=Name
 lbl.namespace=Namespace
 lbl.pattern=Pattern
@@ -182,6 +185,7 @@ page.title.collections=Collections
 page.title.corporate_bodies=Corporate bodies
 page.title.digital_objects=Digital Objects
 page.title.digitization_projects=(Digitization) Projects
+page.title.fileresources=Files
 page.title.geolocations=Geolocations
 page.title.identifier_types=Identifier types
 page.title.items=Items
