+++ dc-cudami-admin/src/main/resources/messages_de.properties
@@ -117,12 +117,15 @@ lbl.collection.inline=Sammlung
 lbl.created=erstellt
 lbl.description=Kurzbeschreibung
 lbl.digital_object.inline=das Digitalisat (Digitale Objekt)
+lbl.filename=Dateiname
+lbl.filesize=Dateigröße
 lbl.firstname=Vorname
 lbl.identifier_type=Identifikatortyp
 lbl.itemlocale=Sprachspezifische Werte anzeigen für
 lbl.label=Bezeichnung
 lbl.last_modified=letzte Änderung
 lbl.lastname=Nachname
+lbl.mimetype=Mimetype
 lbl.name=Name
 lbl.namespace=Namensraum
 lbl.pattern=Muster
@@ -182,6 +185,7 @@ page.title.collections=Sammlungen
 page.title.corporate_bodies=Institutionen
 page.title.digital_objects=Digitalisate (digitale Objekte)
 page.title.digitization_projects=(Digitalisierungs-)Projekte
+page.title.fileresources=Dateien
 page.title.geolocations=Orte
 page.title.identifier_types=Identifikatortypen
 page.title.items=Exemplare
