+++ dc-cudami-admin/src/main/resources/templates/fileresources/list.html
@@ -11,25 +11,155 @@
     <section layout:fragment="content">
       <div class="content-header">
         <div class="container-fluid">
+
           <div class="row">
-            <div class="col-12">
-              <div id="list"></div>
-              <script th:src="@{/js/identifiable-list.bundle.js}"></script>
-              <script th:inline="javascript">
-                IdentifiableList({
-                  apiContextPath: /*[[@{/}]]*/ '',
-                  existingLanguages: /*[[${existingLanguages}]]*/ [],
-                  id: "list",
-                  showEdit: true,
-                  showNew: true,
-                  type: "fileResource",
-                  uiLocale: /*[[${#locale.language}]]*/ ''
-                });
-              </script>
+            <div class="col">
+              <h1 th:text="#{page.title.fileresources}">Files</h1>
+              <hr>
             </div>
           </div>
+
+          <div class="row">
+            <div class="col">
+
+              <div id="toolbar-fileresources">
+                <!-- TODO it always contains "und" undefined language? -->
+                <th:block th:if="${not #lists.isEmpty(existingLanguages)}">
+                  <span th:text="#{lbl.itemlocale}">Show language dependent values in</span>&nbsp;
+                  <select name="locales" id="fileresources-locales">
+                    <th:block th:each="locale : ${existingLanguages}">
+                      <option th:value="${locale.language}" th:text="${locale.getDisplayLanguage(#locale)}" th:selected="${#strings.equals(locale.language, defaultLanguage)}">English</option>
+                    </th:block>
+                  </select>
+                </th:block>
+              </div>
+              <table id="table-fileresources"
+                     data-toggle="table"
+                     th:data-url="@{/api/fileresources}"
+                     th:data-locale="${#locale}"
+                     data-ajax-options="ajaxOptions"
+                     data-buttons="buttons"
+                     data-cookie="true"
+                     data-cookie-id-table="table-fileresources"
+                     data-id-field="uuid"
+                     data-pagination="true"
+                     data-pagination-v-align="both"
+                     data-search="true"
+                     data-search-highlight="true"
+                     data-search-on-enter-key="true"
+                     data-server-sort="true"
+                     data-show-search-button="true"
+                     data-show-search-clear-button="true"
+                     data-side-pagination="server"
+                     data-sort-name="label"
+                     data-sort-order="desc"
+                     data-toolbar="#toolbar-fileresources"
+                     data-unique-id="uuid">
+                <thead>
+                  <tr>
+                    <th data-field="index" data-formatter="formatRowNumber" data-halign="right" data-align="right" th:text="#{lbl.row_number}">#</th>
+                    <th data-field="previewImage" data-formatter="formatPreviewImage" data-halign="center" data-align="center"><i class="fas fa-image"></i></th>
+                    <th data-field="label" data-sortable="true" data-formatter="formatLabel" th:text="#{lbl.label}">Label</th>
+                    <th data-field="filename" data-sortable="true" th:text="#{lbl.filename}">Filename</th>
+                    <th data-field="mimeType" data-sortable="true" th:text="#{lbl.mimetype}">Mimetype</th>
+                    <th data-field="sizeInBytes" data-sortable="true" data-formatter="formatBytesize" th:text="#{lbl.filesize}">Filesize</th>
+                    
+                    <th data-field="lastModified" data-sortable="true" data-formatter="renderDateTime" th:text="#{lbl.last_modified}">last modified</th>
+                    <th data-field="created" data-sortable="true" data-formatter="renderDateTime" th:text="#{lbl.created}">created</th>
+                    <th data-field="actions" data-formatter="formatActions" th:text="#{lbl.actions}">Actions</th>
+                  </tr>
+                </thead>
+              </table>
+
+            </div>
+          </div>
+
         </div>
       </div>
     </section>
-  </body>
+
+  <th:block layout:fragment="beforeBodyEnds">
+    <div th:replace="fragments/modals/confirm-yes-no :: confirm-yes-no-dialog"></div>
+
+    <script type="text/javascript">
+      const $table = $('#table-fileresources');
+
+      window.ajaxOptions = {
+        beforeSend: function (jqXHR, settings) {
+          let itemLocale = $("#fileresources-locales").val();
+          settings.url = settings.url + '&itemLocale=' + itemLocale;
+        }
+      };
+
+      $(function () {
+        $("#fileresources-locales").change(function () {
+          $table.bootstrapTable('refresh');
+        });
+      });
+
+      function formatBytesize(value, row, index) {
+        return value;
+      }
+      
+      function formatRowNumber(value, row, index) {
+        return renderRowNumber($table, index);
+      }
+    </script>
+
+    <script type="text/javascript" th:inline="javascript">
+      /*<![CDATA[*/
+      function buttons() {
+        /*[+
+         const baseUrl = [[@{'/fileresources'}]];
+         const textAdd = [[#{btn.add}]];
+         const tooltipAdd = [[#{tooltip.add_new_table_entry}]];
+         +]*/
+        return {
+          btnAdd: {
+            text: textAdd,
+            icon: 'fa-plus',
+            event: function () {
+              location.href = baseUrl + '/new';
+            },
+            attributes: {
+              title: tooltipAdd
+            }
+          }
+        };
+      }
+      
+      function formatActions(value, row) {
+        /*[+
+         const baseUrl = [[@{'/fileresources'}]] + '/' + row.uuid;
+         +]*/
+        return [
+          renderViewAction(baseUrl),
+          renderEditAction(baseUrl)
+        ].join(' ');
+      }
+
+      function formatIdentifiers(value, row) {
+        return renderIdentifiers(value);
+      }
+
+      function formatLabel(value, row) {
+        /*[+
+         const baseUrl = [[@{'/fileresources'}]] + '/' + row.uuid;
+         +]*/
+        const itemLocale = $("#fileresources-locales").val();
+        return renderLabel(baseUrl, value, itemLocale);
+      }
+
+      function formatPreviewImage(value, row) {
+        /*[+
+         const tooltipNoPreviewImage = [[#{tooltip.no_preview_image}]];
+         +]*/
+        let previewImage = row.previewImage;
+        return renderPreviewImage(previewImage, 30, tooltipNoPreviewImage);
+      }
+      /*]]>*/
+    </script>
+    <div th:replace="fragments/bootstrap-table :: data-formatter"></div>
+  </th:block>
+</body>
 </html>
